refactor backend into package
