+++ tests/github_stats_test.py
@@ -1,6 +1,6 @@
 from datetime import timedelta
 
-from busy_beaver.backend import utc_now_minus
+from busy_beaver.backend.tasks import utc_now_minus
 from busy_beaver.github_stats import generate_summary
 from busy_beaver.models import User
